refactor: 📦 reorganize code structure, optimize type file
