+++ packages/faas/README.md
@@ -1,3 +1,23 @@
 # @mincloudx/faas
 
 > A more user-friendly Node.js cloud function development framework.
+
+## start
+
+Compiles TypeScript files to prepare the package for use after installation.
+
+```bash
+pnpm run tsc
+```
+
+Watches TypeScript files for changes and compiles them in real-time.
+
+```bash
+pnpm run watch
+```
+
+Executes tests using Jest in verbose mode.
+
+```bash
+pnpm run test
+```
