+++ packages/faas/tsconfig.json
@@ -10,6 +10,10 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "outDir": "dist/esm",
-    "declarationDir": "dist/esm/types"
+    "declarationDir": "dist/esm/types",
+    "baseUrl": "./",
+    "paths": {
+      "@local/types/*": ["types/*"]
+    }
   }
 }
